Author & Description Fixes
These changes will reflect new code to add an author endpoint to retrieve data. Now, when we search for a book, it should (in theory) populate the first result that comes up.

This will now allow us to use our db.json file as expected, and pre-populate our page with whatever data we want it to have as of right now. Any other requests we make that are specifically API-related will need review, but any fetch requests to localhost:3000/books/ should now be available to us! Woo!

Styling this will come in a future update, but at least we are getting what we anticipate rendering on the DOM.

We may also want to figure out how to take the API calls and separate them then keep the localhost fetches in this, but it's not necessary for now, just looooong. haha.