Removed 'current book'
We do not have a 'current book' class, so this removes unnecessary css for that.

Also slightly styled the book within the highlighted book area to have a cleaner border like the rest of the styling and updated the size of the book to be L instead of the really small image we were originally targeting.